pyproject.toml
[project]
name = "langchain-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.7",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.3",
    "numpy>=2.3.4",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
